# docs/swagger.yaml
definitions:
...
  internal_handlers.AuditLog:
    properties:
      action:
        type: string
      created_at:
        type: string
      details:
        items:
          type: integer
        type: array
      id:
        type: integer
      ip_address:
        type: string
      success:
        type: boolean
      user_agent:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/audit/logs:
    get:
      consumes:
      - application/json
      description: Get audit logs with optional filters
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Start time (ISO 8601)
        in: query
        name: start
        type: string
      - description: End time (ISO 8601)
        in: query
        name: end
        type: string
      - description: Filter by action type
        in: query
        name: action
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by success status
        in: query
        name: success
        type: boolean
      - description: Limit results (default 100, max 500)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.AuditLog'
            type: array
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get audit logs
      tags:
      - audit
...
swagger: "2.0"